third_party/blink/web_tests/platform/generic/fast/dom/HTMLLinkElement/link-crossOrigin-expected.txt

This tests the handling of the link.crossOrigin attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'crossOrigin' in link is true
PASS link.crossOrigin is null
PASS link.setAttribute('crossorigin', ''); link.crossOrigin is "anonymous"
PASS link.removeAttribute('crossorigin'); link.crossOrigin is null
PASS link.setAttribute('crossorigin', 'foo'); link.crossOrigin is "anonymous"
PASS link.setAttribute('crossorigin', 'ANONYMOUS'); link.crossOrigin is "anonymous"
PASS link.setAttribute('crossorigin', 'use-credentials'); link.crossOrigin is "use-credentials"
PASS link.crossOrigin = ''; link.getAttribute('crossorigin') is ""
PASS link.crossOrigin = 'foo'; link.getAttribute('crossorigin') is "foo"
PASS link.crossOrigin = null; link.hasAttribute('crossorigin') is false
PASS link.crossOrigin = 'ANONYMOUS'; link.getAttribute('crossorigin') is "ANONYMOUS"
PASS link.crossOrigin = 'use-credentials'; link.getAttribute('crossorigin') is "use-credentials"
PASS successfullyParsed is true

TEST COMPLETE
